use Long for WhenOccurs and inlines in IO

+++ core/src/main/scala/swaydb/core/util/English.scala
@@ -18,7 +18,7 @@ package swaydb.core.util
 
 private[swaydb] object English {
 
-  def plural(count: Int, word: String): String =
+  def plural(count: Long, word: String): String =
     if (count == 0 || count > 1)
       word + "s"
     else
